Implement draft of perfect circle resolution for the dbg constructior
Perfect circles don't have ends and so the kmers involved in them don't get incorporated into
a DBG during construction at present. This commit implements a draft improvement
that fixes this problem. It starts picking one of the kmers involved in a circle, and just
keeps going forward, until that same kmer is found again.